Cleaned Up code
